+++ src/OpenTelemetry/Trace/TracerProviderSdk.cs
@@ -23,11 +23,6 @@ using Microsoft.Extensions.DependencyInjection;
 using OpenTelemetry.Internal;
 using OpenTelemetry.Resources;
 
-using CallbackHelper = OpenTelemetry.ProviderBuilderServiceCollectionCallbackHelper<
-    OpenTelemetry.Trace.TracerProviderBuilderSdk,
-    OpenTelemetry.Trace.TracerProviderSdk,
-    OpenTelemetry.Trace.TracerProviderBuilderState>;
-
 namespace OpenTelemetry.Trace
 {
     internal sealed class TracerProviderSdk : TracerProvider
@@ -50,8 +45,8 @@ namespace OpenTelemetry.Trace
         {
             Debug.Assert(serviceProvider != null, "serviceProvider was null");
 
-            var state = serviceProvider!.GetRequiredService<TracerProviderBuilderState>();
-            state.RegisterProvider(nameof(TracerProvider), this);
+            var state = serviceProvider!.GetRequiredService<TracerProviderBuilderSdk>();
+            state.RegisterProvider(this);
 
             this.ServiceProvider = serviceProvider!;
 
@@ -63,17 +58,16 @@ namespace OpenTelemetry.Trace
 
             OpenTelemetrySdkEventSource.Log.TracerProviderSdkEvent("Building TracerProvider.");
 
-            CallbackHelper.InvokeRegisteredConfigureStateCallbacks(
-                serviceProvider!,
-                state);
+            var configureProviderBuilders = serviceProvider!.GetServices<IConfigureTracerProviderBuilder>();
+            foreach (var configureProviderBuilder in configureProviderBuilders)
+            {
+                configureProviderBuilder.ConfigureBuilder(serviceProvider!, state);
+            }
 
             StringBuilder processorsAdded = new StringBuilder();
             StringBuilder instrumentationFactoriesAdded = new StringBuilder();
 
-            if (state.SetErrorStatusOnException)
-            {
-                state.EnableErrorStatusOnException();
-            }
+            state.AddExceptionProcessorIfEnabled();
 
             var resourceBuilder = state.ResourceBuilder ?? ResourceBuilder.CreateDefault();
             resourceBuilder.ServiceProvider = serviceProvider;