+++ src/OpenTelemetry/CHANGELOG.md
@@ -2,10 +2,22 @@
 
 ## Unreleased
 
+## 1.4.0-rc.1
+
+Released 2022-Dec-12
+
 * Added dependency injection support in the `ResourceBuilder` class and added
   support for loading environment variables from `IConfiguration` for the
   `AddEnvironmentVariableDetector` extension (Logs)
   ([#3889](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3889))
+
+* Refactored `AddInstrumentation`, `ConfigureServices` and `ConfigureBuilder`
+  APIs into the OpenTelemetry.Extensions.DependencyInjection package and added
+  the `IServiceCollection.AddOpenTelemetry` API
+  ([#3923](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3923))
+
+* Removed `ConfigureResource` on `OpenTelemetryLoggingOptions`
+  ([#3999](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3999))
 
 ## 1.4.0-beta.3